Implemented inifinite data approach.
`Pipeline.idr` only supports doing one computation in a pipeline, which
is not very interesting. By contrast, `ChannelPL.idr` is a bit more
restrictive in terms of what types one is allowed to pass around in the
pipeline, but supports keeping going until a `DONE` is seen. The latter
might be more what we're after.