Added 2-way channel/pipe version of Pipelines.
This was actually done 4½ hours ago and then I went for lunch and got
distracted, and so forgot to commit. Oops...

To circumvent the problem of System.Concurrency's Channels blocking on
`channelPut`, I reused the code from my attempt at implementing Channels
in pure Idris. The result is a somewhat ugly and confusing pile of code,
but it can be reasoned about and (more importantly) it works as
expected: As far as I can tell, each stage is linked correctly and not
bothered by blocks on sending; Receiving obviously blocks, since
otherwise it is difficult to compute anything, but that is expected
behaviour as far as I'm aware.

As to what I got distracted by, see the next commit.