Added a version based on nicer, directional channels.
Since all the second halves of the bi-directional channels/pipes were of
type `PipelineData ()`, i.e. nothing was being sent, I figured their
was probably a nicer way to do things. Hence, `DirectionalChannels`.
Imo, it works really nicely, both in terms of the types and in terms of
the API. I'm particularly fond of the use of dependent pairs to supply
the send/receive functions, but we'll see if Edwin thinks it's a neat
idea.

Having only one thing in the channel type makes for a lot nicer code,
and as far as I can tell, this implementation is both non-blocking
(apart from receive) and hasn't broken any functionality. Neat!