runFarm also returns a ThreadID; workers stop.
Used the `waitOnThreadBundle` function to make `runFarm` return a single
ThreadID which can then be `threadWait`ed on.

Changed `WorkUnit` ending indication from `DONE` to `STOP`, since this
removes a clash with the new `WorkResult` type, which has a `DONE`. This
also makes more sense imo. Also made sure that `loadWork` actually puts
the `STOP` units at the end of the channel, so that the workers actually
get stopped. There is possibly a nicer way to do that (e.g. having a
custom channel that returns something like `Either STOP (TASK t)`), but
that's a discussion for another time.

TODO: Refactor to use `WorkResult` in the return type, so that I can
define `collect`.