Farm now has a result type and a `collect` function.
The workers in a farm now return a `WorkResult a` instead of just an
`a`. This allows us to know when a worker has stopped, and retrieve the
contents from all workers as seen in the `collect` function. However,
there is still some room for improvement: the calls to `assert_total`
are bugging me; it seems like they should be fixable somehow...